#576 move the draft mode btn to navbar
